Proof suport
This commit introduces basic proof logic, inside proof.rs.
The main logic is under the Proof struct, witch is a collection
of hashes and targets to be proven.

Targets: u64 representing the position of a leaf inside the current
Stump. Only leafs can be passed as targets.

Hashes: Internal hases used by proofs, those are all hashes in witch we
cannot calculate by ourselves, and should be provided from an external
source.

All the verification logic is provided, along with a wide range of tests