Refactor: Optimize roots_to_destroy
`roots_to_destroy` uses a non-zero hash as a marker, holding a root position
that doesn't exist. We accomplish this by calling Hash::from_hex() and
0x01 as a 32-byte vector. However, from_hex involves expensive string
parsing. After some testing, we found this
function using more than 60% of cpu time while updating an accumulator.

This commit solves it by introducing a lazy global variable that
contains this hash already, so we don't have to compute it over and over