Drop redundant dependency
We depend on bitcoin and bitcoin-hashes. Since we only need
bitcoin::hashes from bitcoin, and bitcoin::hashes is a
reexport of bitcoin-hashes, we don't need bitcoin.
Furthermore, depending on bitcoin without specify features pulls libsecp
and some other crates we don't use, but create problems for consumers if
one of the libs try to use a different version of bitcoin.
This commit drops it and use only bitcoin-hashes in its latest version.