Allow Pollard to Verify proofs
While implementing the verify method, the overall way that verify is
handle was changed. Instead of verify being a proof's method, it now is
exposed by every type of accumulator. They all use a common low-level
api from proof, but we need this because every acc handles roots
differently. Moreover, asking for roots and leaves at API level is
verbose and undesirable.