Bugfix: Fix a memleak with pollard
Before this commit we use Rc cycles to represent the forest's trees, but
this creates some floating Rc's that never gets dropped, yelding a
memory leak. This commit fixes this by using Weak references everywhere
except for a node's children, since the node is meant to be owned by
it's ancestor.

It is ok to upgrade those Weak references because they will never
outlive the node itself.