feat: make the hash type generic
Currently, we only support using our internal hash type, that is the
32-bytes representation of a sha512-256 hash digest. However, we may
want to give callers the possibility to use other hash types, or store
them in other ways. This commit modifies `Pollard`, `Proof` and `Stump`
to take in a generic parameter `Hash` that implements the trait
`NodeHash` and defaults to `BitcoinNodeHash`, the one used by Bitcoin
consensus as defined by the utreexo spec.

This is part of a project to support a Cairo prover for Bitcoin. To
reduce the circuit size, we need to use an algebraic hash function like
Poseidon, which tends to reduce the circuit significantly. With this
commit the caller can use our data structures with Poseidon without
needing to change anything in rustreexo.