fix a bug with MemForest due to off-by-one error
When we try to find in what subtree our node is, we loop over all
existing roots. However, this loop is not inclusive, and will never
reach the highest tree. If we don't find the tree, it continues with `0`
as the tree position, but this is wrong, and will cause problems
downstream when we try to use that position.

This commit updates the loop and make it inclusive