Pollard: add a leaf map
When proving, we currently ask for positions. However, positions may
change during updates, so externally holding those positions may be
cumbersome. The leaf hash, on the other hand never changes.

This commit does the same we do for `MemForest`, creating a map from
leaf hashes to a reference to the actual leaf in memory. Once we need to
prove it, we compute the position on-the-fly.