Pollard: add a from_roots
One use for a partial pollard is for mempool and wallet implementations, where you need to
keep track of a subset of the whole tree. Instead of remputing the pollard from genesis for
every flavor of accumulator you have, you can sync-up using the [Stump] and then use its
roots to create a up-to-date pollard.