stump: remove the last calls to sort
This is the last commit in a series that removes unecessary calls to
`sort` that were taking a lot of CPU time.

This removes calls to sort inside detwin, using the same aproach
we've been using of binary searching and inserting where we need it.

It also removes calls to `sort` and `dedup` inside `Stump::add`, by
using a BTreeSet.