Pollard: add proper error handling
We currently only return some strings as error type. This is awful for
API consumers that needs to keep parsing strings in order to undertand
what is going on. This commit introduces a `PollardError` enum that is
used with all result types from Pollard operations.

This provides better error handling and recovery for our users.